Separate empty child
