+++ web/html/src/manager/content-management/shared/components/panels/sources/channels/empty-child.tsx
@@ -1,0 +1,13 @@
+import * as React from "react";
+
+import styles from "./channels-selection.css";
+
+const EmptyChild = () => {
+  return (
+    <div className={styles.nested_row}>
+      <span>&nbsp;{t("no child channels")}</span>
+    </div>
+  );
+};
+
+export default EmptyChild;
